final socket auth config
