+++ server/bin/www.js
@@ -38,7 +38,6 @@ const io = require("socket.io")(server);
  */
 
 const authenticate = async (socket, data, callback) => {
-  console.log(onlineUsers);
   const username = data.username;
   try {
     const user = await User.findOne({
@@ -64,7 +63,7 @@ io.on("connection", (socket) => {
   /**
    * Lookup online user with socket id
    */
-  
+
   socket.on("new-message", (data) => {
     socket.to(onlineUsers[data.recipientId]).emit("new-message", {
       message: data.message,
@@ -75,15 +74,15 @@ io.on("connection", (socket) => {
 
   socket.on("logout", (id) => {
     if (onlineUsers[id]) {
-      console.log("logout");
+      socket.disconnect();
       delete onlineUsers[id];
     }
-    console.log("failed loguot");
   });
-
-  socket.onAny((event, ...args) => {
-    console.log(event, args);
-  });
+  // Uncomment for socket logging
+  
+  // socket.onAny((event, ...args) => {
+  //   console.log(event, args);
+  // });
 });
 
 auth(io, {
