+++ server/package.json
@@ -21,7 +21,7 @@
     "nodemon": "^2.0.6",
     "pg": "^8.5.1",
     "sequelize": "^6.3.5",
-    "socket.io": "^3.0.4",
+    "socket.io": "^3.1.2",
     "socketio-auth": "^0.1.1"
   },
   "devDependencies": {
